If $XDG_CONFIG_DIRS is already set, don't discard it.
Some upstreams like LXQt will want to get configuration from locations
that aren't /etc/xdg{,/xdg-$DESKTOP_SESSION}. This ensures that if
$XDG_CONFIG_DIRS is already set, while it isn't put at the beginning of
the variable (so correct settings are still set), it makes sure it's
still included.

Not doing this is also against the upstream XDG spec, where it states:
"If $XDG_CONFIG_DIRS is either not set or empty, a value equal to
/etc/xdg should be used."
The key word being "if" - if $XDG_CONFIG_DIRS is indeed set, we need to
include that as well.